Use the proper formula for the random seed. The coordinates are multiplied as ints and are cast to longs for the final addition.

+++ src/main/java/com/github/lunatrius/core/world/chunk/ChunkHelper.java
@@ -8,9 +8,9 @@ public class ChunkHelper {
     private static final Random RANDOM = new Random();
 
     public static boolean isSlimeChunk(final long seed, final BlockPos pos) {
-        final long x = pos.getX() >> 4;
-        final long z = pos.getZ() >> 4;
-        RANDOM.setSeed(seed + (x * x * 4987142L) + (x * 5947611L) + (z * z * 4392871L) + (z * 389711L) ^ 987234911L);
+        final int x = pos.getX() >> 4;
+        final int z = pos.getZ() >> 4;
+        RANDOM.setSeed(seed + (long) (x * x * 4987142) + (long) (x * 5947611) + (long) (z * z) * 4392871L + (long) (z * 389711) ^ 987234911L);
         return RANDOM.nextInt(10) == 0;
     }
 }
